Move eventbase definition into event.hpp
