+++ rmw_zenoh_cpp/src/detail/rmw_data_types.hpp
@@ -31,100 +31,14 @@
 
 #include "rcutils/allocator.h"
 
-#include "rmw/event_callback_type.h"
 #include "rmw/rmw.h"
 
+#include "event.hpp"
 #include "graph_cache.hpp"
 #include "message_type_support.hpp"
 #include "service_type_support.hpp"
 
 /// Structs for various type erased data fields.
-
-/// A struct to store status changes which can be mapped to rmw event statuses.
-/// The data field can be used to store serialized information for more complex statuses.
-struct rmw_zenoh_event_status_t
-{
-  size_t total_count;
-  size_t total_count_change;
-  size_t current_count;
-  std::string data;
-};
-
-///=============================================================================
-/// Base class to be inherited by entities that support events.
-class EventsBase
-{
-public:
-  /// @brief Set the user defined callback that should be called when
-  /// a new message/response/request is received.
-  /// @param user_data the data that should be passed to the callback.
-  /// @param callback the callback to be set.
-  void set_user_callback(const void * user_data, rmw_event_callback_t callback);
-
-  /// Trigger the user callback.
-  void trigger_user_callback();
-
-  /// @brief  Set the callback to be triggered when the relevant event is triggered.
-  /// @param event_id the id of the event
-  /// @param callback the callback to trigger for this event.
-  /// @param user_data the data to be passed to the callback.
-  void event_set_callback(
-    rmw_zenoh_event_type_t event_id,
-    rmw_event_callback_t callback,
-    const void * user_data);
-
-  /// @brief  Returns true if the event queue is empty.
-  /// @param event_id the event id whose event queue should be checked.
-  bool event_queue_is_empty(rmw_zenoh_event_type_t event_id) const;
-
-  /// Pop the next event in the queue.
-  /// @param event_id the event id whose queue should be popped.
-  std::unique_ptr<rmw_zenoh_event_status_t> pop_next_event(
-    rmw_zenoh_event_type_t event_id);
-
-  /// Add an event status for an event.
-  /// @param event_id the event id queue to which the status should be added.
-  void add_new_event(
-    rmw_zenoh_event_type_t event_id,
-    std::unique_ptr<rmw_zenoh_event_status_t> event);
-
-  /// @brief Attach the condition variable provided by rmw_wait.
-  /// @param condition_variable to attach.
-  void attach_event_condition(
-    rmw_zenoh_event_type_t event_id,
-    std::condition_variable * condition_variable);
-
-  /// @brief Detach the condition variable provided by rmw_wait.
-  void detach_event_condition(rmw_zenoh_event_type_t event_id);
-
-private:
-  /// @brief Trigger the callback for an event.
-  /// @param event_id the event id whose callback should be triggered.
-  void trigger_event_callback(rmw_zenoh_event_type_t event_id);
-
-  /// Notify once event is added to an event queue.
-  void notify_event(rmw_zenoh_event_type_t event_id);
-
-  /// Mutex to lock when read/writing members.
-  // The mutex is recursive as add_new_event() invokes `trigger_event_callback()`.
-  mutable std::recursive_mutex event_mutex_;
-  /// Mutex to lock for event_condition.
-  mutable std::mutex event_condition_mutex_;
-  /// Condition variable to attach for event notifications.
-  std::condition_variable * event_conditions_[ZENOH_EVENT_ID_MAX + 1]{nullptr};
-  /// User callback that can be set via set_user_callback().
-  rmw_event_callback_t callback_ {nullptr};
-  /// User data that should be passed to the user callback.
-  const void * user_data_ {nullptr};
-  /// Count for
-  size_t unread_count_ {0};
-  rmw_event_callback_t event_callback_[ZENOH_EVENT_ID_MAX + 1] {nullptr};
-  const void * event_data_[ZENOH_EVENT_ID_MAX + 1] {nullptr};
-  size_t event_unread_count_[ZENOH_EVENT_ID_MAX + 1] {0};
-  // A dequeue of events for each type of event this RMW supports.
-  std::deque<std::unique_ptr<rmw_zenoh_event_status_t>> event_queues_[ZENOH_EVENT_ID_MAX + 1] {};
-  const std::size_t event_queue_depth_ = 10;
-};
 
 ///=============================================================================
 struct rmw_context_impl_s
